refactor: 이벤트 마커 훅으로 분리 (#64)
