+++ src/components/event/event-map/index.tsx
@@ -10,12 +10,10 @@ import {
   RoundedButton,
   DeferredLoader,
 } from '@/components';
-import { useMyLocationStore, useMapStore } from '@/stores';
+import { useMapStore, useMyLocationStore } from '@/stores';
 import { confirm, getMarker, getCurrentPosition } from '@/utils';
 import { ROUTES } from '@/constants/routes';
-import { useEventFilter } from '@/hooks';
-import { EventData } from '@/types/event';
-import { theme } from '@/styles/theme';
+import { useEventFilter, useMapEventMarkers } from '@/hooks';
 
 window.navermap_authFailure = function () {
   console.error('네이버 지도 인증 실패');
@@ -24,80 +22,20 @@ window.navermap_authFailure = function () {
 
 const EventMap = () => {
   // localStorage.clear();
-  const { selectedEvent, setSelectedEvent } = useMapStore();
-  const { myLocation, setMyLocation } = useMyLocationStore();
-  const { sortedEvents } = useEventFilter();
   const [mapInstance, setMapInstance] = useState<naver.maps.Map>();
-  const [markers] = useState<Map<bigint, naver.maps.Marker>>(
-    new Map<bigint, naver.maps.Marker>(),
-  );
-  const [infoWindows] = useState<Map<bigint, naver.maps.InfoWindow>>(
-    new Map<bigint, naver.maps.InfoWindow>(),
-  );
   const [searchQuery] = useQueryState('event-search', { defaultValue: '' });
   const [isLoading, setIsLoading] = useState(false);
-  const navigate = useNavigate();
-  const location = useLocation();
 
-  // 마커 클릭 이벤트
-  const handleMarkerClick = useCallback(
-    (mapEvent: EventData) => {
-      setSelectedEvent(mapEvent);
-      const marker = markers.get(mapEvent.eventId);
-      const infoWindow = infoWindows.get(mapEvent.eventId);
-      if (!marker || !infoWindow || !mapInstance) return;
-
-      if (infoWindow.getMap()) {
-        // InfoWindow가 열려 있다면 닫기
-        infoWindow.close();
-        setSelectedEvent(null);
-      } else {
-        // InfoWindow가 닫혀 있다면 열기
-        infoWindow.open(mapInstance, marker);
-        // 지도 이동 및 확대
-        mapInstance?.morph(
-          new naver.maps.LatLng(mapEvent.latitude, mapEvent.longitude),
-          19,
-          { duration: 0, easing: 'easeOutCubic' }, // duration은 500ms 고정
-        );
-      }
-    },
-    [setSelectedEvent, markers, infoWindows, mapInstance],
+  const { selectedEvent } = useMapStore();
+  const { myLocation, setMyLocation } = useMyLocationStore();
+  const { sortedEvents } = useEventFilter();
+  const { markers, createMarkers, updateMarkers } = useMapEventMarkers(
+    mapInstance,
+    sortedEvents,
   );
 
-  const showMarker = (map: naver.maps.Map, marker: naver.maps.Marker) => {
-    // 지도에 표시되어있는지 확인
-    if (marker.getMap()) return;
-    // 표시되어있지 않다면 오버레이를 지도에 추가
-    marker.setMap(map);
-  };
-
-  const hideMarker = (marker: naver.maps.Marker) => {
-    // 지도에 표시되어있는지 확인
-    if (!marker.getMap()) return;
-    // 표시되어있다면 오버레이를 지도에서 삭제
-    marker.setMap(null);
-  };
-
-  const updateMarkers = useCallback(() => {
-    if (!mapInstance) return;
-
-    // 현재 지도의 화면 영역을 mapBounds 변수에 저장
-    const mapBounds = mapInstance.getBounds();
-
-    // 마커 객체 배열을 순회하며 각 마커의 위치를 position 변수에 저장
-    markers.forEach((marker) => {
-      const position = marker.getPosition();
-
-      if (mapBounds.hasPoint(position)) {
-        // 마커가 화면에 보이면 표시
-        showMarker(mapInstance, marker);
-      } else {
-        // 화면 밖이면 숨김
-        hideMarker(marker);
-      }
-    });
-  }, [markers, mapInstance]);
+  const navigate = useNavigate();
+  const location = useLocation();
 
   const initMap = useCallback(
     (centerLat: number, centerLng: number) => {
@@ -129,80 +67,10 @@ const EventMap = () => {
       });
       markers.set(0n, myLocMarker);
 
-      // 이벤트 마커들
-      sortedEvents.forEach((event: EventData) => {
-        const markerOptions = {
-          position: new naver.maps.LatLng(event.latitude, event.longitude),
-          map: mapInstance as naver.maps.Map,
-          icon: {
-            content: getMarker(event.category.name),
-            size: new naver.maps.Size(36, 36),
-            origin: new naver.maps.Point(0, 0),
-            anchor: new naver.maps.Point(18, 18),
-          },
-        };
-        const marker = new naver.maps.Marker(markerOptions);
-        markers.set(event.eventId, marker); // 각 이벤트 마커들은 event id로 구분
-
-        // InfoWindow
-        const InfoWindowContent = document.createElement('div');
-        InfoWindowContent.innerHTML = `
-          <div class="select-speech-bubble" style="position: relative;">
-            <div class="selected-content" style="
-              position: absolute;
-              top: -40px;
-              right: 50%;
-              transform: translateX(50%);
-              border-radius: ${theme.borderRadius.sm};
-              color: ${theme.color.gray[0]};
-              background: ${theme.color.gray[900]};
-              display: flex;
-              align-items: center;
-              justify-content: center;
-              padding: 7px 10px;
-              box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-            ">
-              <div class="title" style="
-                max-width: 200px;
-                ${theme.typeFace.body['15B']};
-                text-overflow: ellipsis;
-                overflow: hidden;
-                white-space: nowrap;
-              ">${event.title}</div>
-            </div>
-            <div class="speech-bubble-tail" style="
-              position: absolute;
-              bottom: -5px;
-              left: 50%;
-              transform: translateX(-50%);
-              width: 16px;
-              height: 10px;
-              background: ${theme.color.gray[900]};
-              clip-path: polygon(0 0, 100% 0, 50% 100%);
-              box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-            "/>
-          </div>
-        `;
-        const infowindow = new naver.maps.InfoWindow({
-          content: InfoWindowContent,
-          // 기본 설정 무시
-          borderWidth: 0,
-          disableAnchor: true,
-          backgroundColor: 'transparent',
-          // 위치
-          pixelOffset: new naver.maps.Point(0, -35),
-        });
-        infoWindows.set(event.eventId, infowindow);
-
-        // 이벤트 마커 클릭 이벤트
-        naver.maps.Event.addListener(marker, 'click', () =>
-          handleMarkerClick(event),
-        );
-
-        // console.log('markers', markers);
-      });
+      // 이벤트 마커 생성
+      createMarkers();
     },
-    [mapInstance, handleMarkerClick, markers, sortedEvents, infoWindows],
+    [mapInstance, createMarkers, markers],
   );
 
   // console.log('sortedEvents', sortedEvents);
@@ -254,7 +122,7 @@ const EventMap = () => {
     return () => {
       naver.maps.Event.removeListener(MoveEventListner);
     };
-  }, [idleHandler]);
+  }, [idleHandler, mapInstance]);
 
   const handleLocationSuccess = useCallback(
     async (position: GeolocationPosition) => {
