+++ src/hooks/index.ts
@@ -1,4 +1,10 @@
-export { default as useEventFilter } from './event/useEventFilter';
 export { default as useRecentSearch } from './common/useRecentSearch';
 export { default as useId } from './common/useId';
+
+// event
+export { default as useEventFilter } from './event/useEventFilter';
+// export { default as useMapCluster } from './event/useMapCluster';
+export { default as useMapEventMarkers } from './event/useMapEventMarkers';
+
+// community
 export { default as useCommunityId } from './community/useCommunityId';
